+++ conf/evolutions/default/1.sql
@@ -4,7 +4,7 @@
 # --- !Ups
 
 create table product (
-  id                            bigint not null,
+  id                            integer not null,
   name                          varchar(255),
   category                      varchar(255),
   description                   varchar(255),
